Fixing pattern match consistencies
let ["a", a] = ["a", 1] now works as a result of removing the top-level
special casing, and these two match statements now produce the same
result (although the code generated is different!).

```muse
match (1, 2) {
  (1, 2) => true,
};

let a = (1, 2);
match a {
  (1, 2) => true,
}
```

Both of these examples produced pattern mismatches before, although in
one case Nil was returned due to a bug in the compiler that has also
been fixed.