Self-hosted test suite
This starts moving some types of tests to a muse-executed test suite.

This suite executor has benefits over the rsn one because the rsn one
can't track the actual line numbers to report errors when working on the
test suite or changes that are breaking tests. The new runner is able to
accurately map the line numbers, allowing Ctrl+Click in VS Code to just
work. This goes as far as reporting stack traces of exceptions.

The runner needs some more features to be able to support all of what
the rsn runner currently does, but dogfooding the language with the test
suite is a clear win.