Removing tuples/arrow functions
Following the path of typescript, Muse wasn't really distinguishing
between list and tuple. After fixing pattern match inconsistencies, we
were left with lists and tuples being identical.

After removing the "list of expressions" option for the parentheses
parsing led to an issue: arrow functions converted a tuple to an
argument list, and using a list in that position feels weird. Since we
have so many ways to define a function, losing arrow functions didn't
feel that consequential.