Yielding to the collector
The Vm now has a setting for how many steps to count per each budget
charge, and each budget charge is when we also yield to the garbage
collector. This yielding slowed execution down.

One area where extra collection was happening that didn't need to was
how the registered code was being stored as a weak reference. By
upgrading it to a root, we can avoid constantly dropping the root count
to 0 when calling Function::call(). Roots dropping to 0 will notify the
collector to run, which is the overhead we really are avoiding.

This ended up being a net swap for performance in the fibonacci vm
example.