WIP: Multi-threaded spawnable runtime
This is a decent work in progress but I've come to realize it isn't
going to solve my need for the project I'm working on, and it has its
own technical challenges to solve before being finished.

Future self: The issue with the spawn unit test is that the function
references a local function in the current module. The function is
evaluated dynamically so when an attempt to find a function named
recurse_spawn, it can't find it because the spawned Vm doesn't have the
same modules.

It might be as simple as: the spawning VM's root module is cloned. But
I'm not sure that's quite correct either, it's probably more nuanced
given that the root module could be calling another module that is
creating a closure. Maybe this means the entire VM's state should be
cloned when spawning, but the frames should be reset after cloning.